extensions/instances/bria/requirements.txt
torch == 2.9.1
torchvision === 0.24.1
transformers == 4.57.3
scikit-image === 0.26.0
pillow == 12.0.0
picteus-internal-extension-sdk
